refresh 202409041901

+++ backend/JbdGjsService/00.システム機能/008.マスタメンテナンス/GJ8030_日本養鶏協会マスタメンテナンス/WebInterface/VM.vb
@@ -172,7 +172,7 @@ Namespace JBD.GJS.Service.GJ8030
         ''' <summary>
         ''' 支払　依頼人コード
         ''' </summary>
-        Public Property KOFU_KAISYA_CD As Integer? = Nothing
+        Public Property KOFU_KAISYA_CD As Long? = Nothing
 
         ''' <summary>
         ''' 支払　振込依頼人名（ﾌﾘｶﾞﾅ）
